refactor: merge types
